chore(engine): move version checking functions to a new module "helpers"
